Extract getActionItems to hook and update to make request on page load
